Mute unmute login and register and name sync
all these things working fine
open apps fixed for android..
room detection is buggy and needs some changes

+++ Assets/Agora-RTC-Plugin/API-Example/Tools/RequestToken.cs
@@ -16,7 +16,7 @@ namespace Agora_RTC_Plugin.API_Example
 
     public static class HelperClass
     {
-        public static IEnumerator FetchToken(string url, string channel, int userId, Action<string> callback = null)
+        public static IEnumerator FetchToken(string url, string channel, long userId, Action<string> callback = null)
         {
             using (UnityWebRequest request = UnityWebRequest.Get(string.Format("{0}?channel={1}&userId={2}", url, channel, userId)))
             {
